renamed boltonoptions to bootstrapperoptions
