+++ tests/BoltOn.Tests/Data/EF/Fakes/EFRegistrationTask.cs
@@ -9,21 +9,21 @@ namespace BoltOn.Tests.Data.EF.Fakes
 {
 	public static class EFRegistrationTask
 	{
-		public static void RegisterDataFakes(this BoltOnOptions boltOnOptions)
+		public static void RegisterDataFakes(this BootstrapperOptions bootstrapperOptions)
 		{
-			boltOnOptions.ServiceCollection.AddTransient<IRepository<Student>, Repository<Student, SchoolDbContext>>();
-			boltOnOptions.ServiceCollection.AddTransient<IQueryRepository<Student>, QueryRepository<Student, SchoolDbContext>>();
+			bootstrapperOptions.ServiceCollection.AddTransient<IRepository<Student>, Repository<Student, SchoolDbContext>>();
+			bootstrapperOptions.ServiceCollection.AddTransient<IQueryRepository<Student>, QueryRepository<Student, SchoolDbContext>>();
 
 			if (IntegrationTestHelper.IsSqlServer)
 			{
-				boltOnOptions.ServiceCollection.AddDbContext<SchoolDbContext>(options =>
+				bootstrapperOptions.ServiceCollection.AddDbContext<SchoolDbContext>(options =>
 				{
 					options.UseSqlServer("Data Source=127.0.0.1;initial catalog=Testing;persist security info=True;User ID=sa;Password=Password1;");
 				});
 			}
 			else
 			{
-				boltOnOptions.ServiceCollection.AddDbContext<SchoolDbContext>(options =>
+				bootstrapperOptions.ServiceCollection.AddDbContext<SchoolDbContext>(options =>
 				{
 					options.UseInMemoryDatabase("InMemoryDbForTesting");
 				});
