+++ tests/BoltOn.Tests/Requestor/Fakes/RequestorRegistrationTask.cs
@@ -11,35 +11,35 @@ namespace BoltOn.Tests.Requestor.Fakes
 {
 	public static class RequestorRegistrationTask
 	{
-		public static void RegisterRequestorFakes(this BoltOnOptions boltOnOptions)
+		public static void RegisterRequestorFakes(this BootstrapperOptions bootstrapperOptions)
 		{
 			var appClock = new Mock<IAppClock>();
 			var currentDateTime = DateTime.Parse("10/27/2018 12:51:59 PM");
 			appClock.Setup(s => s.Now).Returns(currentDateTime);
-			boltOnOptions.ServiceCollection.AddTransient((s) => appClock.Object);
+			bootstrapperOptions.ServiceCollection.AddTransient((s) => appClock.Object);
 
 			var testInterceptorLogger = new Mock<IAppLogger<TestInterceptor>>();
 			testInterceptorLogger.Setup(s => s.Debug(It.IsAny<string>()))
 								.Callback<string>(st => RequestorTestHelper.LoggerStatements.Add(st));
-			boltOnOptions.ServiceCollection.AddTransient((s) => testInterceptorLogger.Object);
+			bootstrapperOptions.ServiceCollection.AddTransient((s) => testInterceptorLogger.Object);
 
 			var stopWatchInterceptorLogger = new Mock<IAppLogger<StopwatchInterceptor>>();
 			stopWatchInterceptorLogger.Setup(s => s.Debug(It.IsAny<string>()))
 									 .Callback<string>(st => RequestorTestHelper.LoggerStatements.Add(st));
-			boltOnOptions.ServiceCollection.AddTransient((s) => stopWatchInterceptorLogger.Object);
+			bootstrapperOptions.ServiceCollection.AddTransient((s) => stopWatchInterceptorLogger.Object);
 
 			var transactionInterceptorLogger = new Mock<IAppLogger<TransactionInterceptor>>();
 			transactionInterceptorLogger.Setup(s => s.Debug(It.IsAny<string>()))
 									 .Callback<string>(st => RequestorTestHelper.LoggerStatements.Add(st));
-			boltOnOptions.ServiceCollection.AddTransient((s) => transactionInterceptorLogger.Object);
+			bootstrapperOptions.ServiceCollection.AddTransient((s) => transactionInterceptorLogger.Object);
 
 			if (RequestorTestHelper.IsClearInterceptors)
-				boltOnOptions.RemoveAllInterceptors();
+				bootstrapperOptions.RemoveAllInterceptors();
 
 			if (RequestorTestHelper.IsRemoveStopwatchInterceptor)
-				boltOnOptions.RemoveInterceptor<StopwatchInterceptor>();
+				bootstrapperOptions.RemoveInterceptor<StopwatchInterceptor>();
 
-			boltOnOptions.AddInterceptor<TestInterceptor>();
+			bootstrapperOptions.AddInterceptor<TestInterceptor>();
 		}
 	}
 }
